feat(nvim): migrate to lazy.nvim
