--- modules/nvim/lua/noice-config.lua
@@ -1,33 +1,0 @@
--- ---------------------------------------- noice --------------------------------------------------------
-
-local M = {}
-
-M.load = function()
-  require("noice").setup({
-    messages = {
-      view_search = false,
-    },
-    lsp = {
-      hover = {
-        enabled = false,
-      },
-      progress = {
-        enabled = false,
-      },
-      override = {
-        ["vim.lsp.util.convert_input_to_markdown_lines"] = true,
-        ["vim.lsp.util.stylize_markdown"] = true,
-        ["cmp.entry.get_documentation"] = true,
-      },
-    },
-    presets = {
-      bottom_search = true,
-      command_palette = true,
-      long_message_to_split = true,
-      inc_rename = false,
-      lsp_doc_border = false,
-    },
-  })
-end
-
-return M
