+++ modules/nvim/lua/plugins/lsp/init.lua
@@ -1,0 +1,6 @@
+return {
+  require("plugins.lsp.mason"),
+  require("plugins.lsp.null-ls"),
+  require("plugins.lsp.nvim-dap"),
+  require("plugins.lsp.nvim-lspconfig"),
+}
