+++ modules/nvim/lua/plugins/noice.lua
@@ -1,0 +1,43 @@
+-- ---------------------------------------- noice --------------------------------------------------------
+return {
+  "folke/noice.nvim",
+  dependencies = {
+    "MunifTanjim/nui.nvim",
+    {
+      "rcarriga/nvim-notify",
+      config = function()
+        require("notify").setup({
+          background_colour = "#1e1e2e",
+        })
+        vim.notify = require("notify")
+      end,
+    },
+  },
+  config = function()
+    require("noice").setup({
+      messages = {
+        view_search = false,
+      },
+      lsp = {
+        hover = {
+          enabled = false,
+        },
+        progress = {
+          enabled = false,
+        },
+        override = {
+          ["vim.lsp.util.convert_input_to_markdown_lines"] = true,
+          ["vim.lsp.util.stylize_markdown"] = true,
+          ["cmp.entry.get_documentation"] = true,
+        },
+      },
+      presets = {
+        bottom_search = true,
+        command_palette = true,
+        long_message_to_split = true,
+        inc_rename = false,
+        lsp_doc_border = false,
+      },
+    })
+  end,
+}
