+++ modules/nvim/lua/plugins/snippets.lua
@@ -1,0 +1,6 @@
+-- ---------------------------------------- snippets --------------------------------------------------------
+return {
+  "hrsh7th/vim-vsnip",
+  "hrsh7th/vim-vsnip-integ",
+  "rafamadriz/friendly-snippets",
+}
