+++ modules/nvim/lua/plugins/vimtex.lua
@@ -1,0 +1,16 @@
+-- ---------------------------------------- vimtex --------------------------------------------------------
+
+return {
+  {
+    "lervag/vimtex",
+    config = function()
+      vim.g.vimtex_indent_enabled = 1
+      vim.g.vimtex_indent_conditionals = {}
+      vim.g.vimtex_indent_on_ampersands = 0
+      vim.g.vimtex_complete_close_braces = 1
+      vim.g.vimtex_format_enabled = 1
+      vim.g.vimtex_imaps_leader = ";"
+      vim.g.vimtex_quickfix_open_on_warning = 0
+    end,
+  },
+}
