--- modules/nvim/lua/treesitter-config.lua
@@ -1,32 +1,0 @@
--- ---------------------------------------- treesitter ---------------------------------------------------------
-
-local M = {}
-
-M.load = function()
-  require("nvim-treesitter.configs").setup({
-    ensure_installed = "all",
-    highlight = {
-      enable = true,
-    },
-    incremental_selection = {
-      enable = false,
-    },
-    indent = {
-      enable = true,
-    },
-    -- does not compile on macOS
-    ignore_install = {
-      "phpdoc",
-    },
-  })
-
-  -- workaround issue with vim-markdown
-  vim.api.nvim_exec(
-    [[
-    au BufNewFile,BufRead *.md TSBufDisable highlight
-    ]],
-    false
-  )
-end
-
-return M
